+++ JAVA4All/src/org/Java4All/DOM/ReadXMLDOM.java
@@ -1,3 +1,4 @@
+package org.Java4All.DOM;
 import java.io.IOException;
 
 import javax.xml.parsers.DocumentBuilder;
