+++ JAVA4All/src/org/Java4All/Helloworld/Hello.java
@@ -1,3 +1,4 @@
+package org.Java4All.Helloworld;
 
 public class Hello {
      public static void main(String[] args)
