+++ JAVA4All/src/org/Java4All/LinkedList/linkedListSen.java
@@ -1,4 +1,6 @@
-package com.senvel.Helper;
+package org.Java4All.LinkedList;
+
+
 
 public class linkedListSen {
 	private Node headNode;
