+++ JAVA4All/src/org/Java4All/WordCounter/WordCounter.java
@@ -1,0 +1,8 @@
+package org.Java4All.WordCounter;
+
+
+public class WordCounter {
+	public static void main (String args[]){
+		System.out.println("Word Counter Program starts");
+	}
+}
