+++ JAVA4All/src/org/Java4All/http/ReadUrl.java
@@ -1,3 +1,4 @@
+package org.Java4All.http;
 import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
